Upgrade 3.4.3 (#13)
* [#12] Checking that applying our config works

jboss-cli rollback the configuration if any line fails, but the script still returns 0.

Greping the results for a result that is different from { "output": "success"}

* [#12] Being explicit about when to apply the non-cache for themes

So that the dev configuration is not applied when running in CI

* [#12] Testing that the login page has the custom Akvo css

* [#12] Updating Keycloak version

Updating some of the configuration that no longer works in 3.4

* [#12] Removing deprecation warning on JDBC_PING

* [#12] Removing deprecation warning on pbcast.GMS

* [#12] Removing deprecation warning on FD_SOCK

* [#12] Upgrade FD to FD_ALL as per migration script

+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM jboss/keycloak:3.1.0.Final
+FROM jboss/keycloak:3.4.3.Final
 
 ENV PROXY_ADDRESS_FORWARDING=true \
     JB_HOME=/opt/jboss \
@@ -8,7 +8,8 @@ COPY docker-entrypoint.sh ${JB_HOME}
 COPY theme/akvo/ ${KC_HOME}/themes/akvo/
 COPY keycloak_configuration.txt ${KC_HOME}
 
-RUN ${KC_HOME}/bin/jboss-cli.sh --file=${KC_HOME}/keycloak_configuration.txt && \
+RUN ${KC_HOME}/bin/jboss-cli.sh --file=${KC_HOME}/keycloak_configuration.txt | tee ${KC_HOME}/cli-result.txt && \
+    grep outcome ${KC_HOME}/cli-result.txt | grep -v success && exit 123 || \
     rm -rf ${KC_HOME}/standalone/configuration/standalone_xml_history/current/* && \
     mkdir -p ${KC_HOME}/modules/system/layers/base/com/mysql/jdbc/main; \
     cd ${KC_HOME}/modules/system/layers/base/com/mysql/jdbc/main && \
